Rename useList to useGridList
