+++ packages/@react-aria/gridlist/README.md
@@ -1,3 +1,3 @@
-# @react-aria/list
+# @react-aria/gridlist
 
 This package is part of [react-spectrum](https://github.com/adobe/react-spectrum). See the repo for more details.
