+++ packages/@react-aria/gridlist/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@react-aria/list",
+  "name": "@react-aria/gridlist",
   "version": "3.0.0-beta.0",
   "description": "Spectrum UI components in React",
   "license": "Apache-2.0",
