+++ packages/@react-spectrum/list/package.json
@@ -34,9 +34,9 @@
     "@babel/runtime": "^7.6.2",
     "@react-aria/button": "^3.5.1",
     "@react-aria/focus": "^3.6.1",
+    "@react-aria/gridlist": "3.0.0-beta.0",
     "@react-aria/i18n": "^3.4.1",
     "@react-aria/interactions": "^3.9.1",
-    "@react-aria/list": "3.0.0-beta.0",
     "@react-aria/utils": "^3.13.1",
     "@react-aria/virtualizer": "^3.4.1",
     "@react-aria/visually-hidden": "^3.3.1",
@@ -46,15 +46,12 @@
     "@react-spectrum/progress": "^3.2.1",
     "@react-spectrum/text": "^3.2.1",
     "@react-spectrum/utils": "^3.7.1",
-    "@react-stately/collections": "^3.4.1",
     "@react-stately/layout": "^3.5.0",
     "@react-stately/list": "^3.5.1",
     "@react-stately/virtualizer": "^3.2.1",
-    "@react-types/grid": "^3.1.1",
     "@react-types/list": "3.0.0-alpha.1",
     "@react-types/shared": "^3.13.1",
-    "@spectrum-icons/ui": "^3.3.0",
-    "react-transition-group": "^4.4.2"
+    "@spectrum-icons/ui": "^3.3.0"
   },
   "devDependencies": {
     "@adobe/spectrum-css-temp": "^3.0.0-alpha.1",
