+++ packages/@react-spectrum/list/src/index.ts
@@ -1,3 +1,3 @@
 export {ListView} from './ListView';
 export {Item} from '@react-stately/collections';
-export type {SpectrumListProps} from '@react-types/list';
+export type {SpectrumListViewProps as SpectrumListProps} from '@react-types/list';
